requirements.txt
streamlit==1.28.0
google-cloud-bigquery==3.11.4
pandas-gbq==0.19.2
pandas==2.0.3
numpy==1.24.3